requirements.txt
aiogram>=3.0.0
yt-dlp>=2024.0.0
motor>=3.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
dnspython>=2.4.0